Move accessibilityUtils under app/utils dir
